Alter keys of Shopping List Item
I propose that we simplify the key structure for the shopping list item. If we make the `_id` a simple `item:cj6mn7e36000001p9n14fgk6s` value and add a `list` which is the `_id` of the parent shopping list, then we are able to extract the data cleanly with Mango e.g.

```
{
  selector: { 
     type: 'item',
     list: 'list:cj6mj1zfj000001n1ugjfkj33'
   }
}
```

This is much easier to understand than key mangling on the primary index and allows us to sorting by a third value, for instance.